+++ Mahmoud/Fetching-and-Updating-Data-with-React-Query/react-course-part2-starter-main/src/react-query/hooks/useTodos.ts
@@ -1,0 +1,22 @@
+import { useQuery } from "@tanstack/react-query";
+import axios from "axios";
+
+interface Todo {
+  id: number;
+  title: string;
+  userId: number;
+  completed: boolean;
+}
+
+const useTodos = () => {
+  const fetchTodos = () =>
+    axios
+      .get<Todo[]>("https://jsonplaceholder.typicode.com/todos")
+      .then((res) => res.data);
+  return useQuery<Todo[], Error>({
+    queryKey: ["todos"],
+    queryFn: fetchTodos,
+  });
+};
+
+export default useTodos;
